requirements.txt
requests
SQLAlchemy
python-dotenv
boto3
snowflake-sqlalchemy
Jinja2
gensim
seaborn
folium
pandas
numpy